+++ languagetool-core/src/main/java/org/languagetool/rules/Rule.java
@@ -48,7 +48,7 @@ public abstract class Rule {
 
   protected final ResourceBundle messages;
 
-  private List<String> correctExamples = new ArrayList<>();
+  private List<CorrectExample> correctExamples = new ArrayList<>();
   private List<IncorrectExample> incorrectExamples = new ArrayList<>();
   private ITSIssueType locQualityIssueType = ITSIssueType.Uncategorized;
   private Category category;
@@ -191,14 +191,14 @@ public abstract class Rule {
   /**
    * Set the examples that are correct and thus do not trigger the rule.
    */
-  public final void setCorrectExamples(List<String> correctExamples) {
+  public final void setCorrectExamples(List<CorrectExample> correctExamples) {
     this.correctExamples = Objects.requireNonNull(correctExamples);
   }
 
   /**
    * Get example sentences that are correct and thus will not match this rule.
    */
-  public final List<String> getCorrectExamples() {
+  public final List<CorrectExample> getCorrectExamples() {
     return Collections.unmodifiableList(correctExamples);
   }
 
@@ -300,7 +300,7 @@ public abstract class Rule {
    * with the error corrected.
    * @since 2.5
    */
-  protected void addExamplePair(IncorrectExample incorrectSentence, String correctSentence) {
+  protected void addExamplePair(IncorrectExample incorrectSentence, CorrectExample correctSentence) {
     incorrectExamples.add(incorrectSentence);
     correctExamples.add(correctSentence);
   }